requirements.txt
# 核心依赖
Flask>=2.3.0
Flask-CORS>=4.0.0
Pillow>=9.5.0

# 可选依赖（如果需要更多功能）
requests>=2.31.0 